Support for no-target-directory and hidden files
add option for no-target-director, analagous to `cp`'s
no-target-directory. Expected behavior is that when copying a directory
to another directory, instead of creating a sub-folder in target,
overwrite target.

add option for hidden files. Default is to ignore hidden files.

This is a little hacked in; it seems as though `ignore`'s `Walk` could
handle both .gitignore and hidden files on its own, but I may have
missed something.